Add admin endpoint integration tests; fix overall_score serialization
Round 11. With local test capability from Round 10, cover the
previously untested /api/admin/* endpoints and fix a
serialization mismatch surfaced by those tests.

Tests (tests/test_admin.py, 8 TestClient cases)
- stats on empty db, trace list/detail/404, agent filter, golden
  label filter (JSONB extras), turn-job trace ordering, game
  evaluations, and admin token auth.

Fix
- admin.TurnEvaluationRead.overall_score: Decimal -> float. FastAPI
  serialized the Numeric(3,2) column as the JSON string "4.17", which
  disagreed with the frontend lib/api.ts type number. Returning float
  makes the JSON a real number.

Verified locally against pgvector pg16: full suite 77 passed. Also
confirmed the turn_evaluations FK constraint works (inserting a fake
turn_id was rejected; the test now builds a real game + turn).

+++ api/app/routers/admin.py
@@ -7,7 +7,6 @@
 from __future__ import annotations
 
 from datetime import datetime
-from decimal import Decimal
 from typing import Any
 from uuid import UUID
 
@@ -133,7 +132,10 @@ class TurnEvaluationRead(BaseModel):
     prose_quality: int | None
     freshness: int | None
     safety: int | None
-    overall_score: Decimal | None
+    # DB 列是 Numeric(3,2)（Decimal）；用 float 让 JSON 序列化为 number 而非 string，
+    # 与前端 TurnEvaluationRead.overall_score: number 一致。Pydantic v2 lax 模式
+    # 会把 Decimal coerce 成 float。
+    overall_score: float | None
     rationale: dict[str, Any] | None
     judge_model: str | None
     trace_id: UUID | None
